+++ vault/resource_aws_auth_backend_client.go
@@ -101,6 +101,12 @@ func awsAuthBackendClientResource() *schema.Resource {
 				Computed:    true,
 				Description: "The TTL of generated identity tokens in seconds.",
 			},
+			consts.FieldMaxRetries: {
+				Type:        schema.TypeInt,
+				Default:     -1,
+				Optional:    true,
+				Description: "Number of max retries the client should use for recoverable errors.",
+			},
 		},
 	}
 }
@@ -119,11 +125,10 @@ func awsAuthBackendWrite(ctx context.Context, d *schema.ResourceData, meta inter
 	stsEndpoint := d.Get(consts.FieldSTSEndpoint).(string)
 	stsRegion := d.Get(consts.FieldSTSRegion).(string)
 	stsRegionFromClient := d.Get(useSTSRegionFromClient).(bool)
-
 	identityTokenAud := d.Get(consts.FieldIdentityTokenAudience).(string)
 	roleArn := d.Get(consts.FieldRoleArn).(string)
 	identityTokenTTL := d.Get(consts.FieldIdentityTokenTTL).(int)
-
+	maxRetries := d.Get(consts.FieldMaxRetries).(int)
 	iamServerIDHeaderValue := d.Get(consts.FieldIAMServerIDHeaderValue).(string)
 
 	path := awsAuthBackendClientPath(backend)
@@ -134,6 +139,7 @@ func awsAuthBackendWrite(ctx context.Context, d *schema.ResourceData, meta inter
 		consts.FieldSTSEndpoint:            stsEndpoint,
 		consts.FieldSTSRegion:              stsRegion,
 		consts.FieldIAMServerIDHeaderValue: iamServerIDHeaderValue,
+		consts.FieldMaxRetries:             maxRetries,
 	}
 
 	if d.HasChange(consts.FieldAccessKey) || d.HasChange(consts.FieldSecretKey) {
@@ -204,6 +210,7 @@ func awsAuthBackendRead(ctx context.Context, d *schema.ResourceData, meta interf
 		consts.FieldSTSEndpoint,
 		consts.FieldSTSRegion,
 		consts.FieldIAMServerIDHeaderValue,
+		consts.FieldMaxRetries,
 	}
 	for _, k := range fields {
 		if v, ok := secret.Data[k]; ok {